+++ src/users.h
@@ -45,15 +45,15 @@ public:
     QVariant data(const QModelIndex &index, int role = TypeRole) const;
     QHash<int, QByteArray> roleNames() const;
 
-    std::shared_ptr<User> getUser(int32_t userId) const;
-    Q_INVOKABLE QVariant getUserAsVariant(int32_t userId) const;
+    std::shared_ptr<User> getUser(int64_t userId) const;
+    Q_INVOKABLE QVariant getUserAsVariant(int64_t userId) const;
 
 public slots:
     void updateUser(td_api::updateUser *updateUser);
     void onUpdateUserFullInfo(td_api::updateUserFullInfo* updateUserFullInfo);
 
 private:
-    QHash<int32_t, std::shared_ptr<User>> _users;
+    QHash<int64_t, std::shared_ptr<User>> _users;
     std::shared_ptr<TelegramManager> _manager;
     shared_ptr<Files> _files;
